refactor(architecture): move some generators to `core-workflow-steps`
