+++ antenna-workflow-steps/pom.xml
@@ -36,9 +36,5 @@
         <module>processors/validators/antenna-security-issue-validator</module>
         <module>processors/validators/antenna-license-validator</module>
         <module>processors/validators/antenna-match-state-validator</module>
-        <!-- generators -->
-        <module>generators/antenna-csv-generator</module>
-        <module>generators/antenna-HTML-report-generator</module>
-        <module>generators/antenna-source-zip-generator</module>
     </modules>
 </project>
