+++ antenna/core-workflow-steps/src/main/java/org/eclipse/sw360/antenna/workflow/generators/CSVGenerator.java
@@ -20,7 +20,7 @@ import org.eclipse.sw360.antenna.model.artifact.facts.java.BundleCoordinates;
 import org.eclipse.sw360.antenna.model.artifact.facts.java.MavenCoordinates;
 import org.eclipse.sw360.antenna.model.util.ArtifactLicenseUtils;
 import org.eclipse.sw360.antenna.model.xml.generated.LicenseInformation;
-import org.eclipse.sw360.antenna.repository.Attachable;
+import org.eclipse.sw360.antenna.api.Attachable;
 
 import java.io.File;
 import java.io.FileNotFoundException;
