+++ antenna/core-workflow-steps/src/main/resources/licenseReport.vm
