+++ antenna/model/src/main/java/org/eclipse/sw360/antenna/api/Attachable.java
@@ -9,7 +9,7 @@
  * SPDX-License-Identifier: EPL-2.0
  */
 
-package org.eclipse.sw360.antenna.repository;
+package org.eclipse.sw360.antenna.api;
 
 import org.eclipse.sw360.antenna.api.IAttachable;
 import org.eclipse.sw360.antenna.api.exceptions.AntennaExecutionException;
